chore: organized getSummary in a file
