+++ web/src/App.tsx
@@ -2,29 +2,13 @@ import { Dialog } from './components/ui/dialog'
 import { CreateGoal } from './components/create-goal'
 import { Summary } from './components/summary'
 import { EmptyGoals } from './components/empty-goals'
+import { getSummary } from './http/get-summary'
 import { useQuery } from '@tanstack/react-query'
 
-type SummaryResponse = {
-  completed: number
-  total: number
-  goalsPerDay: Record<
-    string,
-    {
-      id: number
-      title: string
-      completedAt: Date
-    }[]
-  >
-}
-
 export function App() {
-  const { data } = useQuery<SummaryResponse>({
+  const { data } = useQuery({
     queryKey: ['summary'],
-    queryFn: async () => {
-      const response = await fetch('http://localhost:3333/summary')
-      const data = await response.json()
-      return data.summary
-    },
+    queryFn: getSummary,
   })
 
   return (
